tagging IPAv4 4.2.2

+++ ipalib/plugins/baseldap.py
@@ -1565,6 +1565,8 @@ class LDAPDelete(LDAPMultiQuery):
 
     has_output_params = global_output_params
 
+    subtree_delete = True
+
     def execute(self, *keys, **options):
         ldap = self.obj.backend
 
@@ -1600,6 +1602,8 @@ class LDAPDelete(LDAPMultiQuery):
             except errors.NotFound:
                 self.obj.handle_not_found(*nkeys)
             except errors.NotAllowedOnNonLeaf:
+                if not self.subtree_delete:
+                    raise
                 # this entry is not a leaf entry, delete all child nodes
                 delete_subtree(dn)
 
@@ -1918,14 +1922,14 @@ class LDAPSearch(BaseLDAPCommand, crud.Search):
     takes_options = (
         Int('timelimit?',
             label=_('Time Limit'),
-            doc=_('Time limit of search in seconds'),
+            doc=_('Time limit of search in seconds (0 is unlimited)'),
             flags=['no_display'],
             minvalue=0,
             autofill=False,
         ),
         Int('sizelimit?',
             label=_('Size Limit'),
-            doc=_('Maximum number of entries returned'),
+            doc=_('Maximum number of entries returned (0 is unlimited)'),
             flags=['no_display'],
             minvalue=0,
             autofill=False,
